Fix FOVScanner (sort of).
Use an inner loop inside the targets loop to account for the fact that there may be several components of a given type on a target.

Remove the check that there is nothing between the origin and the target. Obviously this does not make it a true FOV, but for the purposes of detected whether a meele attack will connect it will work fine. I had trouble getting it to work, so until there is need for true FOV it is not worth the effotrt.

#22